NEW: Added collection functions API bindings. Some more functions to add here but first they need to check permissions internally (many rely on the page to do the permissions check).
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@8665 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/collections_functions.php
@@ -356,6 +356,9 @@ function delete_collection($ref)
 	# Deletes the collection with reference $ref
 	global $home_dash;
 	
+	# Permissions check
+	if (!collection_writeable($ref)) {return false;}
+	
 	hook("beforedeletecollection","",array($ref));
 	sql_query("delete from collection where ref='$ref'");
 	sql_query("delete from collection_resource where collection='$ref'");
